Move timer to statusBar component
Moved timer and related methods to statusBar component, including pause, close, and closeModal.
