+++ src/components/statusBar/statusBar.js
@@ -1,17 +1,67 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { Progress } from 'semantic-ui-react';
+import { Progress, Button, Modal } from 'semantic-ui-react';
 import { isMobile } from 'react-device-detect';
 import StatusBarStyles from '../styles/StatusBarStyles';
+import { startQuiz, closeQuiz } from '../../actions/quizActions';
+import msToTime from '../../helpers/msToTime';
+import TimerStyles from '../styles/TimerStyles';
 
 class StatusBar extends Component {
   constructor() {
     super();
-    
-    this.state = {};
+
+    this.state = {
+      open: false,
+      time: 0,
+      timerOn: false,
+    };
+
+    this.pause = this.pause.bind(this);
+    this.closeModal = this.closeModal.bind(this);
+    this.close = this.close.bind(this);
   }
+
+  componentDidMount() {
+    this.setState({ timerOn: true, time: 0 });
+    const x = Date.now();
+    this.timer = setInterval(
+      () => this.setState({ time: Date.now() - x }),
+      1000
+    );
+  }
+
+  pause() {
+    const { timerOn } = this.state;
+    if (timerOn) {
+      clearInterval(this.timer);
+      this.setState({ timerOn: false, open: true });
+    }
+  }
+
+  close() {
+    const { closeQuiz } = this.props;
+    closeQuiz();
+    clearInterval(this.timer);
+    this.setState({ time: 0, timerOn: false });
+  }
+
+  closeModal() {
+    const { timerOn, time } = this.state;
+    if (!timerOn) {
+      this.setState({ timerOn: true, open: false }, () => {
+        const x = Date.now() - time;
+        this.timer = setInterval(
+          () => this.setState({ time: Date.now() - x }),
+          1000
+        );
+      });
+    }
+  }
+
   render() {
     const { quiz, quizGuesses, quizAnswers } = this.props.quiz;
+    const { time, open } = this.state;
     const percentComp = quiz
       ? parseInt((quizGuesses.length / quizAnswers.length) * 100, 10)
       : '';
@@ -19,10 +69,38 @@ class StatusBar extends Component {
       quizAnswers.length
     }`;
     const scoreText = `Score: ${quizGuesses.filter(x => x).length}`;
+    const pauseStyle =
+      quizGuesses.length === quizAnswers.length ? { display: 'none' } : {};
+
+    if (quizGuesses.length === quizAnswers.length) {
+      clearInterval(this.timer);
+    }
 
     return (
       <div>
         <StatusBarStyles quiz={quiz} mobile={isMobile}>
+          <TimerStyles mobile={isMobile}>
+            <div className="statusBar-timerButtons">
+              <Button
+                size="mini"
+                compact
+                inverted
+                color="red"
+                className="statusBar-stop"
+                icon="stop"
+                onClick={this.close}
+              />
+              <Button
+                size="mini"
+                compact
+                inverted
+                color="blue"
+                icon="pause"
+                onClick={this.pause}
+                style={pauseStyle}
+              />
+            </div>
+          </TimerStyles>
           <Progress
             percent={percentComp}
             className="statusBar-progress"
@@ -31,8 +109,25 @@ class StatusBar extends Component {
           <div className="statusBar-ratio">
             <p>{questionText}</p>
             <p>{scoreText}</p>
+            <p className="statusBar-timer">{msToTime(time)}</p>
           </div>
         </StatusBarStyles>
+        <Modal
+          basic
+          dimmer="blurring"
+          open={open}
+          onClose={this.closeModal}
+          closeOnDimmerClick={false}
+          style={{ textAlign: 'center' }}
+        >
+          <Button
+            inverted
+            color="green"
+            size="massive"
+            content="Resume"
+            onClick={this.closeModal}
+          />
+        </Modal>
       </div>
     );
   }
@@ -42,4 +137,7 @@ const mapStateToProps = state => ({
   quiz: state.quiz,
 });
 
-export default connect(mapStateToProps)(StatusBar);
+export default connect(
+  mapStateToProps,
+  { startQuiz, closeQuiz }
+)(StatusBar);
